Update minSDK to 24 for latest ZXing
Newer ZXing versions won't work on API levels below 24 because
ZXing requires support for Java 8 now. See;
https://github.com/zxing/zxing/issues/1170

Because this is a major leap forward, this commit also migrates
to AndroidX (since the support library is deprecated) and removes
everything that isn't required anymore.

Development for API levels 9 to 23 will continue in the master
branch. I'm not going to drop support for API levels below 24!